+++ BE/src/chat/chat.controller.ts
@@ -8,10 +8,10 @@ import {
   UseGuards,
 } from '@nestjs/common';
 import { ChatService } from './chat.service';
-import { AuthGuard } from '../utils/auth.guard';
-import { UserHash } from '../utils/auth.decorator';
+import { AuthGuard } from '../common/guard/auth.guard';
+import { UserHash } from '../common/decorator/auth.decorator';
 import { CreateRoomDto } from './createRoom.dto';
-import { FcmHandler } from '../utils/fcmHandler';
+import { FcmHandler } from '../common/fcmHandler';
 
 @Controller('chat')
 export class ChatController {
