+++ BE/src/chat/chat.module.ts
@@ -5,7 +5,7 @@ import { ChatsGateway } from './chats.gateway';
 import { TypeOrmModule } from '@nestjs/typeorm';
 import { ChatRoomEntity } from '../entities/chatRoom.entity';
 import { PostEntity } from '../entities/post.entity';
-import { FcmHandler } from '../utils/fcmHandler';
+import { FcmHandler } from '../common/fcmHandler';
 import { RegistrationTokenEntity } from '../entities/registrationToken.entity';
 import { ChatEntity } from 'src/entities/chat.entity';
 import { UserEntity } from 'src/entities/user.entity';
