+++ BE/src/common/greenEyeHandler.ts
