+++ BE/src/common/hashMaker.ts
