+++ BE/src/post/post.controller.ts
@@ -18,11 +18,11 @@ import { ApiOperation, ApiTags } from '@nestjs/swagger';
 import { UpdatePostDto } from './dto/postUpdate.dto';
 import { FilesInterceptor } from '@nestjs/platform-express';
 import { PostCreateDto } from './dto/postCreate.dto';
-import { MultiPartBody } from '../utils/multiPartBody.decorator';
+import { MultiPartBody } from '../common/decorator/multiPartBody.decorator';
 import { PostListDto } from './dto/postList.dto';
-import { AuthGuard } from 'src/utils/auth.guard';
-import { UserHash } from 'src/utils/auth.decorator';
-import { FilesSizeValidator } from '../utils/files.validator';
+import { AuthGuard } from 'src/common/guard/auth.guard';
+import { UserHash } from 'src/common/decorator/auth.decorator';
+import { FilesSizeValidator } from '../common/files.validator';
 
 @Controller('posts')
 @ApiTags('posts')
