+++ BE/src/post/post.module.ts
@@ -3,11 +3,11 @@ import { PostController } from './post.controller';
 import { PostService } from './post.service';
 import { TypeOrmModule } from '@nestjs/typeorm';
 import { PostEntity } from '../entities/post.entity';
-import { S3Handler } from '../utils/S3Handler';
+import { S3Handler } from '../common/S3Handler';
 import { PostImageEntity } from '../entities/postImage.entity';
 import { BlockUserEntity } from '../entities/blockUser.entity';
 import { BlockPostEntity } from '../entities/blockPost.entity';
-import { AuthGuard } from 'src/utils/auth.guard';
+import { AuthGuard } from 'src/common/guard/auth.guard';
 
 @Module({
   imports: [
