+++ BE/src/posts-block/posts-block.controller.ts
@@ -8,8 +8,8 @@ import {
   UseGuards,
 } from '@nestjs/common';
 import { PostsBlockService } from './posts-block.service';
-import { AuthGuard } from 'src/utils/auth.guard';
-import { UserHash } from 'src/utils/auth.decorator';
+import { AuthGuard } from 'src/common/guard/auth.guard';
+import { UserHash } from 'src/common/decorator/auth.decorator';
 
 @Controller('posts/block')
 @UseGuards(AuthGuard)
