+++ BE/src/report/report.controller.ts
@@ -1,8 +1,8 @@
 import { Body, Controller, Post, UseGuards } from '@nestjs/common';
 import { CreateReportDto } from './createReport.dto';
 import { ReportService } from './report.service';
-import { AuthGuard } from '../utils/auth.guard';
-import { UserHash } from '../utils/auth.decorator';
+import { AuthGuard } from '../common/guard/auth.guard';
+import { UserHash } from '../common/decorator/auth.decorator';
 
 @Controller('report')
 @UseGuards(AuthGuard)
