+++ BE/src/users-block/users-block.module.ts
@@ -4,7 +4,7 @@ import { UsersBlockController } from './users-block.controller';
 import { TypeOrmModule } from '@nestjs/typeorm';
 import { BlockUserEntity } from 'src/entities/blockUser.entity';
 import { UserEntity } from 'src/entities/user.entity';
-import { AuthGuard } from 'src/utils/auth.guard';
+import { AuthGuard } from 'src/common/guard/auth.guard';
 
 @Module({
   imports: [TypeOrmModule.forFeature([BlockUserEntity, UserEntity])],
