+++ BE/src/users/users.module.ts
@@ -2,16 +2,16 @@ import { Module } from '@nestjs/common';
 import { UsersService } from './users.service';
 import { UsersController } from './users.controller';
 import { TypeOrmModule } from '@nestjs/typeorm';
-import { S3Handler } from 'src/utils/S3Handler';
+import { S3Handler } from 'src/common/S3Handler';
 import { UserEntity } from '../entities/user.entity';
 import { PostEntity } from '../entities/post.entity';
 import { PostImageEntity } from '../entities/postImage.entity';
 import { BlockUserEntity } from '../entities/blockUser.entity';
 import { BlockPostEntity } from '../entities/blockPost.entity';
-import { AuthGuard } from 'src/utils/auth.guard';
+import { AuthGuard } from 'src/common/guard/auth.guard';
 import { RegistrationTokenEntity } from '../entities/registrationToken.entity';
-import { FcmHandler } from 'src/utils/fcmHandler';
-import { GreenEyeHandler } from '../utils/greenEyeHandler';
+import { FcmHandler } from 'src/common/fcmHandler';
+import { GreenEyeHandler } from '../common/greenEyeHandler';
 
 @Module({
   imports: [
